refactor(snake.py): Game gives Snake movement dict
Game gives Snake the movement dict minus the initial waiting movement (called None), and the initial snake length. Snake randomly chooses a movement and grows the snake in that direction until the initial snake length is matched. It returns the movement so that Game knows how to set the last move.